requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==7.4.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
requests==2.31.0
urllib3==2.0.6